+++ cmd/baton-freshservice/config.go
@@ -5,18 +5,19 @@ import (
 	"github.com/spf13/viper"
 )
 
-var (
-	// ConfigurationFields defines the external configuration required for the
-	// connector to run. Note: these fields can be marked as optional or
-	// required.
-	ConfigurationFields = []field.SchemaField{}
+const apiKey = "api-key"
+const domain = "domain"
 
-	// FieldRelationships defines relationships between the fields listed in
-	// ConfigurationFields that can be automatically validated. For example, a
-	// username and password can be required together, or an access token can be
-	// marked as mutually exclusive from the username password pair.
-	FieldRelationships = []field.SchemaFieldRelationship{}
+var (
+	apiKeyField         = field.StringField(apiKey, field.WithRequired(true), field.WithDescription("The api key for your account."))
+	domainField         = field.StringField(domain, field.WithRequired(true), field.WithDescription("The domain for your account."))
+	categoryField       = field.StringField("category-id", field.WithDescription("The category id to filter service items to"))
+	configurationFields = []field.SchemaField{apiKeyField, domainField, categoryField}
 )
+
+var configRelations = []field.SchemaFieldRelationship{
+	field.FieldsDependentOn([]field.SchemaField{categoryField}, []field.SchemaField{field.ListTicketSchemasField}),
+}
 
 // ValidateConfig is run after the configuration is loaded, and should return an
 // error if it isn't valid. Implementing this function is optional, it only
